Fixed breaking of EC connection when IP filter reload/update/change-prio was triggered from amulegui
What happened was:
- core added a task to the thread scheduler
- this called CMuleThread::Stop() and the Wait() call there
- the Wait() set up a new event queue instance
- this started processing the next EC package or the rest of the current one, ending up with a zlib error or an empty packet assertion

Now the IP filter calls get posted into an event and are processed after the EC package.
Created a new set of DoNotifyAlways functions to post events regardless if current thread is main or background.
These should be used for all EC commands that created scheduler tasks.